ComCon

Description

A library for running actions in parallel on multiple processor cores.

Documentation

pforever :: (Enum a, Num a) => a -> IO b1 -> IO b

run action on the given number of cores concurrently (blocks: all cores have to finish before next actions are taken)

pforever' :: Int -> IO a -> IO ()

run the action on the given number of cores forever (not blocking)

pfor :: Int -> [IO a] -> IO ()

run the actions on the given number of cores If a core finished an action it picks the next (if there is any left) thus all cores will be busy as long as there are actions left.

pforM_ :: Int -> [a1] -> (a1 -> IO a) -> IO ()

apply function before running pfor

pforM_ p xs f = pfor p (f <$> xs)