Compare

Description

Generating results in CSV format.

Documentation

unkomma :: [Char] -> [Char]

Replaces ',' by '-' (to avoid unwanted separation in CSV format).

values :: Acceptance acc => Formula -> Automat q a l acc -> [Transformer] -> [Text]

Generates a result line in CSV format: formula, tool chain, nr. states, size of acceptance condition

allpathsvalues

Arguments

:: Bool
Apply equivalence check using autfilt

-> [Transformer]

-> Maybe Int
timeout in seconds

-> Maybe Int
maximal length of tool chain

-> Formula

-> IO [[Text]]

Generates result CSV for one formula.

Runs the given transformers and generates result lines for resulting parity automata

allpathsvalues_upto :: Int -> Bool -> [Transformer] -> Maybe Int -> Maybe Int -> Formula -> IO (Bool, [[Text]])

Same as allpathsvalues. Additionally it is computed if the minimal number of states is at least the given number.

rabinreducevalues :: Bool -> [Transformer] -> Maybe Int -> Maybe Int -> Formula -> IO [[Text]]

Same as allpathsvalues but generates result lines for parity automata and Rabin automata.

allpathssuccess :: Bool -> [Transformer] -> Maybe Int -> Maybe Int -> Formula -> IO (Bool, [[Text]])

Same as allpathsvalues. Additionally it is computed if all paths succeeded.

allpathseval

Arguments

:: Bool

-> [Transformer]

-> Maybe Int

-> Maybe Int

-> (Transformer -> Bool)
property

-> Formula

-> IO (Bool, [Text])
(min_in < min_notin, [min_in / min_notin, formula, min_in, t_min_in, min_notin, t_min_notin] )

Compares paths using transformers that have a property (trs_in) to other transformers.

Compares the resulting automata on the minimal number of states computed by tool chains using transformers from trs_in (min_in) and by tool chains not using transformers from trs_in (min_notin).

Checks if one of the tool chains using one of the transformers from trs_in computes a smaller automaton than every tool chain not using transformers from trs_in (min_in < min_notin).

Shows the minimal number of states (min_in, min_notin) and one of the tool chains achieving it (t_min_in, t_min_notin) for each of the two groups of transformers.

Computes the quotient of the minimal number of states (min_in / min_notin).