Paths

Description

Generates all tool chains from given transformers. Translates a LTL formula using these tool chains.

Types

type WayGraph = Gr V Transformer

type for the graph between automata types

type Chain = [Transformer]

data Bunch

results stored in each node of the graph

Constructors

Bunch

Fields

object :: D

the object (formula, automaton)

toolchain :: Chain

the tool chain the object was compouted by

complete :: Bool

have all successors of this node been considered ?

steps_remaining :: Maybe Int

remaining number of steps

type Bunchs = [Bunch]

type State = Map Node Bunchs

While translating a formula in each node the results are stored

Properties

is_done :: State -> Bool

A state is done if all stored bunches are completely processed

can_continue :: Bunch -> Bool

the maximal length of the tool chain has not already been reached

Helper

buildGraph :: [Transformer] -> (WayGraph, Map V Node)

builds the graph from the given transformers

initstate

Arguments

:: WayGraph
graph

-> Maybe Int
maximal lenght of tool chains

-> Node
initial node

-> Set D
initial objects

-> State
initial state

computes the initial state where at the given node the set of objects to start with are stored

testequiv :: FilePath -> D -> ExceptT Failure IO ()

call autfilt --equivalent-to= filepath with the object (automaton) if autfilt fails (exit code 1) this function will fail (special case: autfilt exits with exit code 2 (too many acceptance sets) then this function will not fail)

Translations

step

Arguments

:: (D -> ExceptT Failure IO ())
action for checks on the results (e.g. a call of testequiv) if this check fails, the tool chain is not successfull

-> WayGraph

-> Maybe Int

-> State

-> Node

-> IO State

Computes the direct successors of one node in the graph if a transformer fails (timeout, failure or failure of check) no new result will be stored

state_step :: (D -> ExceptT Failure IO ()) -> WayGraph -> Maybe Int -> State -> IO State

computes the next state by applying step to each node

state_steps :: (D -> ExceptT Failure IO ()) -> WayGraph -> Maybe Int -> State -> IO State

applies state_step until the state is_done

allpaths

Arguments

:: Bool
apply equivalence check using autfilt

-> [Transformer]
the transformers

-> V
VForm

-> D
the formula

-> Maybe Int
timeout

-> Maybe Int
max. length of tool chains

-> IO State
the state after completely processing all tool chains, in each node all (intermediate) results are stored and can be evaulated

translates a formula along the tool chains resulting from the given transformers (optionally calls testequiv with the output of ltl2tgba --generic --deterministic -S on the formula)

allpathswithmap :: Bool -> [Transformer] -> V -> D -> Maybe Int -> Maybe Int -> IO (State, Map V Node)

as allpaths but returns the map from node to automaton type as well