Pathtest

Description

Compares transformers on a given formula or randomly generated formualae.

Documentation

mainwürfel :: ArgOpts -> [Transformer] -> IO ()

compare transformers on randomly generated formulae according to the options

mainformula :: ArgOpts -> [Transformer] -> IO ()

compare transformers on a given formula