Patterns

Description

Compares transformers on a pattern.

Documentation

mainpattern :: ArgOpts -> [Transformer] -> IO ()

Compare transformers on a pattern according to the options. for each path translate the formulae up to the smallest parameter which could not be translated