Transformer

Description

Call of a transformer and translations bewteen different automata types from the literature as transformers.

Constructors

data V

Node types

Constructors

VForm
LTL formula (spot's format)

VSpinForm
LTL formula (spin's format)

VNBA

VDPA

VDRA

VDRAcompact
compact Rabin automaton

VDRAscc
strongly compact Rabin automaton

VDRAsingleton
Rabin automaton where each final set is singleton

data D

data types stored in nodes

Constructors

DForm Formula

DSpinForm SpinFormula

DNBA (NBA State (Maybe LabelExpr) (Maybe LabelExpr))

DDPA (DPA State (Maybe LabelExpr) (Maybe LabelExpr))

DDRA (DRA State (Maybe LabelExpr) (Maybe LabelExpr))

data Failure

Constructors

TO
timeout

PE Text
parse error

Error Text
other error

Transformer

data Transformer

Constructors

Transformer

Fields

source :: V

target :: V

name :: String

exe :: FilePath

args :: [String]

apply

Arguments

:: Transformer

-> D

-> Maybe Int
timeout

-> ExceptT Failure IO D

apply a transformer to an object (formula or automaton) optionally with a timeout

if a timeout is given and the transformer takes too long, exits not successfully or the output could not be parsed this function will fail.

Important: The transformer has to compute an automaton of its target type or else the parser will fail.

Functions

paths :: [Transformer] -> V -> [[Transformer]]

compute all tool chains using the given transformers starting at the given node

pathsbetween :: [Transformer] -> V -> V -> [[Transformer]]

compute all tool chains using the given transformers between the given nodes

Tools From the Literature

LTL -> NBA

spin :: Transformer

ltl2ba :: Transformer

ltl3ba :: Transformer

goal_ltl2nba :: Transformer

ltl2tgba_nba :: Transformer

ltl2nbas :: [Transformer]

Others

ltl2tgba_dpa :: Transformer

piterman :: Transformer

schewe :: Transformer

ltl2dstar :: Transformer

rabinizer :: Transformer

IAR

iar :: Transformer

siar :: Transformer

iaropt :: Transformer

siaropt :: Transformer

transformers :: [Transformer]

all transformers (except rabinreducers)

Reductions

setsplit :: Transformer

setred :: Transformer

setcombine :: Transformer

setsat :: Transformer

setirred :: Transformer

topsplit :: Transformer

topred :: Transformer

topcombine :: Transformer

topsat :: Transformer

topirred :: Transformer

reduceall :: Transformer

rabinreducers :: [Transformer]

either only reduceall or all reductions with the corresponding epsilon transformers

Special

eps :: V -> V -> Transformer

epsilon transformer bewteen the given nodes

ltltospin :: Transformer

transforms a formula in spot's syntax into spin's syntax